Split plugin tests into multiple files
